New improved version with new settings
New improved drop and play pbo for exile servers that spawns in roaming jets. Advise to unpbo and change the config.sqf to your liking then repbo and drop in ExileServer\addons folder. Added setting in config, You can now set a max number of jets to have on the map at one time. Got rid of min and max spawn times there is one spawn time and that is time in seconds between each jet spawn until it reaches the max jet amount. Waypoints have changed, It is scripted to get your map size and generate waypoints randomly. Fixed issue with not being able to use modded jets now you can use any jet 1 or 2 seats and modded Jets. Fixed the Bug of the jets sometimes shooting each other and targeting players in helicopters. They no longer target helicopters or each other. Planes that cannot fight jets are Blacklisted in config to be ignored by the AI jets. You can add to the Blacklist if you have planes you don't want the AI jets attacking. Jets do not attack ground units. There is also a % chance setting for a loot crate to spawn on exposion of AI jet that is shot and destroyed and will parachute to ground displaying chordinates in chat channel. These chordinates are as it spawns so they change slightly as it falls to ground but that makes players have to hunt them and not find them so easily. So in a nutshell your players in helicopters and planes that are blacklisted are safe from attack and this makes it nice for the players who like to fly jets to dog fight other jets.